Move FolderState and OperationType to file_operations
