+++ app/src/main/java/com/amaze/filemanager/ui/activities/MainActivity.java
@@ -28,17 +28,17 @@ import static android.os.Build.VERSION_CODES.KITKAT_WATCH;
 import static android.os.Build.VERSION_CODES.LOLLIPOP;
 import static android.os.Build.VERSION_CODES.M;
 import static android.os.Build.VERSION_CODES.N;
-import static com.amaze.filemanager.filesystem.FolderStateKt.WRITABLE_OR_ON_SDCARD;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.COMPRESS;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.COPY;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.DELETE;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.EXTRACT;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.MOVE;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.NEW_FILE;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.NEW_FOLDER;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.RENAME;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.SAVE_FILE;
-import static com.amaze.filemanager.filesystem.OperationTypeKt.UNDEFINED;
+import static com.amaze.filemanager.file_operations.filesystem.FolderStateKt.WRITABLE_OR_ON_SDCARD;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.COMPRESS;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.COPY;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.DELETE;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.EXTRACT;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.MOVE;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.NEW_FILE;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.NEW_FOLDER;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.RENAME;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.SAVE_FILE;
+import static com.amaze.filemanager.file_operations.filesystem.OperationTypeKt.UNDEFINED;
 import static com.amaze.filemanager.ui.dialogs.SftpConnectDialog.ARG_ADDRESS;
 import static com.amaze.filemanager.ui.dialogs.SftpConnectDialog.ARG_DEFAULT_PATH;
 import static com.amaze.filemanager.ui.dialogs.SftpConnectDialog.ARG_EDIT;
